split app.js into focused modules; state as truth; idempotent re-init
app.js (609 lines at review time) was at least five things. Now:
- setupView.js — the URL setup screen + flashLabel
- updateChecker.js — shell-version probe + force update; reads the
  version.json emitted by the SW codegen instead of regex-parsing the
  live sw.js source (the fragile textual coupling the review flagged)
- statusBar.js — every #update message: loading / freshness / offline /
  storage-full / no-data / error states in one place
- csvExporter.js — the Download Data action
- filterEngine.js — the pure Home filter pipeline (date+part stage
  exposed separately because the Player dropdown populates from it)
App keeps boot, revalidate, wiring, and theme re-render orchestration.

State stops living in the DOM: the Part filter's source of truth is
navigationComponent.selectedPart and the active tab is
tabComponent.activeTab — .part-btn.active/.active-tab classes are now
reflections, never read back.

Re-init (error → re-enter URL → initialize again) is idempotent: menu
document/window listeners are wired once with teardown handles kept,
part buttons and the tab strip rebuild instead of stacking, the calendar
clears .calendar-gen nodes first, and the dashboard re-inits via setData.
